fixed warnings and added nyxstone in benchmark tests

+++ llarm-asm/src/encoding_utils.cpp
@@ -24,8 +24,8 @@ void encoders::encode_imm(u32 &binary, const u32 immed) {
     for (u8 rot = 0; rot < 16; ++rot) {
         u32 rotated = llarm::util::rotr(immed, rot * 2);
         if ((rotated & 0xFFFFFF00) == 0) {
-            llarm::util::swap_bits(binary, 8, 11, static_cast<u8>(rotated));
-            llarm::util::swap_bits(binary, 0, 7, static_cast<u8>(immed));
+            llarm::util::swap_bits(binary, 8, 11, rotated);
+            llarm::util::swap_bits(binary, 0, 7, immed);
         }
     }
 }
